move all scalar funcs
