+++ native/spark-expr/src/lib.rs
@@ -24,11 +24,9 @@ mod error;
 mod checkoverflow;
 pub use checkoverflow::CheckOverflow;
 mod kernels;
-pub mod scalar_funcs;
 mod schema_adapter;
 pub use schema_adapter::SparkSchemaAdapterFactory;
 
-pub mod spark_hash;
 mod normalize_nan;
 
 pub mod test_common;
@@ -38,32 +36,26 @@ pub use unbound::UnboundColumn;
 pub mod utils;
 pub use normalize_nan::NormalizeNaNAndZero;
 
-mod comet_scalar_funcs;
-mod string_funcs;
 mod agg_funcs;
-mod conversion_funcs;
+mod array_funcs;
 mod bitwise_funcs;
+mod comet_scalar_funcs;
+mod conditional_funcs;
+mod conversion_funcs;
+mod datetime_funcs;
+pub mod hash_funcs;
+mod json_funcs;
 mod math_funcs;
 mod predicate_funcs;
-mod conditional_funcs;
-mod array_funcs;
+mod static_invoke;
+mod string_funcs;
 mod struct_funcs;
-mod datetime_funcs;
-mod json_funcs;
 
 pub use comet_scalar_funcs::create_comet_physical_fun;
+use conversion_funcs::*;
 pub use error::{SparkError, SparkResult};
-pub use json_funcs::*;
-pub use string_funcs::*;
-pub use agg_funcs::*;
-pub use conversion_funcs::*;
-pub use bitwise_funcs::*;
-pub use math_funcs::*;
-pub use predicate_funcs::*;
-pub use conditional_funcs::*;
-pub use array_funcs::*;
-pub use struct_funcs::*;
-pub use datetime_funcs::*;
+use math_funcs::*;
+use string_funcs::*;
 
 /// Spark supports three evaluation modes when evaluating expressions, which affect
 /// the behavior when processing input values that are invalid or would result in an
