+++ native/spark-expr/src/math_funcs/internal/mod.rs
@@ -1,0 +1,5 @@
+mod make_decimal;
+mod unscaled_value;
+
+pub use make_decimal::spark_make_decimal;
+pub use unscaled_value::spark_unscaled_value;
