+++ native/spark-expr/src/math_funcs/mod.rs
@@ -1,3 +1,17 @@
+mod ceil;
+mod div;
+mod floor;
+pub(crate) mod hex;
+pub mod internal;
 mod negative;
+mod round;
+pub(crate) mod unhex;
+mod utils;
 
-pub use negative::{NegativeExpr, create_negate_expr};
+pub use ceil::spark_ceil;
+pub use div::spark_decimal_div;
+pub use floor::spark_floor;
+pub use hex::spark_hex;
+pub use negative::{create_negate_expr, NegativeExpr};
+pub use round::spark_round;
+pub use unhex::spark_unhex;
