+++ native/spark-expr/src/predicate_funcs/mod.rs
@@ -1,3 +1,5 @@
+mod is_nan;
 mod rlike;
 
+pub use is_nan::spark_isnan;
 pub use rlike::RLike;
