+++ native/spark-expr/src/static_invoke/char_varchar_utils/mod.rs
@@ -1,0 +1,3 @@
+mod read_side_padding;
+
+pub use read_side_padding::spark_read_side_padding;
