+++ native/spark-expr/src/static_invoke/mod.rs
@@ -1,0 +1,3 @@
+mod char_varchar_utils;
+
+pub use char_varchar_utils::*;
